Usecases created to fetch data (GetFeedUseCase, GetFlightsUseCase and GetTicketsUseCase); SearchMapper for feature:search module (ui layer) created; :data:search module injected in the rest of application;

+++ feature/search/src/main/java/com/livmas/search/domain/dtos/TicketDTO.kt
@@ -12,7 +12,7 @@ data class TicketDTO(
     val startCity: String,
     val endCity: String,
 
-    val timeInTrip: Float,
+    val timeInTrip: Double,
     val hasTransfer: Boolean,
 
     val badge: String? = null
