+++ feature/search/src/main/java/com/livmas/search/ui/flights/adapter/FlightModel.kt
@@ -3,7 +3,7 @@ package com.livmas.search.ui.flights.adapter
 import java.util.Calendar
 
 internal data class FlightModel(
-    val flightId: Int,
+    val flightId: Long,
     val company: String,
     val price: Int,
     val times: List<Calendar>
